% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tpl_logo.R
\name{add_tpl_logo}
\alias{add_tpl_logo}
\title{Texas Policy Lab \link{ggplot2} theme}
\usage{
add_tpl_logo(plot, size = "regular", position = "top", align = 0)
}
\arguments{
\item{plot}{The plot object passed to \code{add_tpl_logo}}

\item{size}{The size of the TPL logo}

\item{position}{The position of the logo. Options are "top", "bottomright", and "bottomleft"}

\item{align}{An optional specification of horizontal nudging of the TPL logo}
}
\description{
Add TPL logo to a single plot as grid object
}
