% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tpl_logo_text.R
\name{add_tpl_logo_text}
\alias{add_tpl_logo_text}
\title{Texas Policy Lab \link{ggplot2} theme}
\usage{
add_tpl_logo_text(plot, align = 0)
}
\description{
Add TPL logo text to a single plot as grid object
}
