% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_tx_cities}
\alias{add_tx_cities}
\title{Add Texas cities}
\usage{
add_tx_cities(label = FALSE)
}
\arguments{
\item{label}{Whether the plotted cities should also have labels}
}
\description{
Add Texas cities
}
