% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_axis.R
\name{drop_axis}
\alias{drop_axis}
\title{Texas Policy Lab \link{ggplot2} theme}
\usage{
drop_axis(axis = "both")
}
\arguments{
\item{axis}{Axis to drop "x", "y", "both", or "neither". The function will keep whatever is not included in "drop."}
}
\description{
Drops tick axes and axes labels
}
