% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_install.R
\name{font_install}
\alias{font_install}
\title{Import and register Adobe Caslon Pro & Lato font}
\usage{
font_install()
}
\description{
\code{font_install} tests to see if default fonts imported and registered. If Adobe Caslon Pro & Lato are not imported and registered, then \code{font_install} imports and registers Adobe Caslon Pro & Lato with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Adobe Caslon Pr must be installed on your computer for \code{font_install} to work.

Test to see if Adobe Caslon Pro is imported and registered with \code{font_test()}.
}
