% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximanova_install.R
\name{proximanova_install}
\alias{proximanova_install}
\title{Import and register Proxima Nova font}
\usage{
proximanova_install()
}
\description{
\code{proximanova_install} tests to see if Proxima Nova is imported and registered. If Proxima Nova isn't imported and registered, then \code{proximanova_install} imports and registers Proxima Nova with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Proxima Nova must be installed on your computer for \code{proximanova_install} to work.

Test to see if Prooxima Nova is imported and registered with \code{proximanova_test()}.
}
