% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximanova_test.R
\name{proximanova_test}
\alias{proximanova_test}
\title{Test for Proxima Nova import and registration}
\usage{
proximanova_test()
}
\description{
\code{proximanova_test} tests to see if Proxima Nova is imported and registered.
}
\details{
Import and register Proxima Nova in R with proximanova_install().
}
