% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_word.R
\name{read_word}
\alias{read_word}
\title{Read word}
\usage{
read_word(pth)
}
\arguments{
\item{pth}{the path to the word document to read}
}
\description{
Reads text from word into Rmarkdown files
}
\seealso{
read_word_table
}
