% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_word.R
\name{read_word_table}
\alias{read_word_table}
\title{Read Word table}
\usage{
read_word_table(pth)
}
\arguments{
\item{pth}{the path to the word document to read}
}
\description{
Reads in a table from word and prints it using the knitr and kableExtra package
}
\examples{
\dontrun{
pth <- "path/table.docx"
df <- read_word_table(pth)
}
}
\seealso{
read_word
}
