% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_ordinal}
\alias{scale_colour_ordinal}
\title{Discrete color scale}
\usage{
scale_colour_ordinal(palette = "categorical", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tpl_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{discrete_scale()}}
}
\description{
Discrete color scale
}
