% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_continuous}
\alias{scale_fill_continuous}
\title{Continuous fill scale}
\usage{
scale_fill_continuous(..., colours = config$palettes$seq[9][[1]],
  colors = config$palettes$seq[9][[1]], values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar")
}
\arguments{
\item{...}{other arguments passed to \code{discrete_scale()}}

\item{colours}{vector of colours}

\item{colors}{vector of colours}

\item{values}{if colours should not be evenly positioned along the gradient this vector gives the position (between 0 and 1) for each colour in the colours vector. See rescale for a convience function to map an arbitrary range to between 0 and 1}

\item{space}{colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.}

\item{na.value}{default color for NA values}

\item{guide}{legend representation of scale}
}
\description{
Continuous fill scale
}
