% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_TPL_print.R
\name{theme_TPL_print}
\alias{theme_TPL_print}
\title{A \link{ggplot2} theme formatted in TPL style}
\usage{
theme_TPL_print(base_size = 11, base_family = "Proxima Nova",
  base_line_size = 0.5, base_rect_size = 0.5)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_TPL} provides a \link{ggplot2} theme formatted according to the TPL website
}
