% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tpl_texas.R
\name{theme_tpl_texas}
\alias{theme_tpl_texas}
\title{A Texas shaped \link{ggplot2} theme formatted in TPL style}
\usage{
theme_tpl_texas(scale = "continuous", base_size = 8.5,
  base_family = "Adobe Caslon Pro", base_line_size = 0.5,
  base_rect_size = 0.5)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_tpl} provides a \link{ggplot2} theme formatted according to the TPL
website
}
