% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tpl_color_pal}
\alias{tpl_color_pal}
\title{Texas Policy Lab [ggplot2] theme}
\usage{
tpl_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used by TPL
}
