% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{tpl_colors}
\alias{tpl_colors}
\title{Function to extract `TPL_color_list` colors as hex codes}
\usage{
tpl_colors(...)
}
\arguments{
\item{...}{Character names of TPL_color_list colors}
}
\description{
Function to extract `TPL_color_list` colors as hex codes
}
