% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_font_install.R
\name{tpl_font_install}
\alias{tpl_font_install}
\title{Import and register Adobe Caslon Pro & Lato font}
\usage{
tpl_font_install()
}
\description{
\code{tpl_font_install} tests to see if default fonts imported and registered. If Adobe Caslon Pro & Lato are not imported and registered, then \code{tpl_font_install} imports and registers Adobe Caslon Pro & Lato with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Adobe Caslon Pro and Lato must be installed on your computer for \code{tpl_font_install} to work.

Test to see if Adobe Caslon Pro and Lato are imported and registered with \code{tpl_font_test()}.
}
