% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_font_test.R
\name{tpl_font_test}
\alias{tpl_font_test}
\title{Test for Adobe Caslon Pro & Lato import and registration}
\usage{
tpl_font_test()
}
\description{
\code{tpl_font_test} tests to see if default fonts are imported and registered.
}
\details{
Import and register Adobe Caslon Pro & Lato in R with tpl_font_install().
}
