% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_TPL_main}
\alias{palette_TPL_main}
\title{Texas Policy Lab color palettes}
\format{An object of class \code{NULL} of length 0.}
\source{
Texas Policy Lab website
  \url{https://www.texaspolicylab.org/}
}
\usage{
palette_TPL_main
}
\description{
Vectors with hex-color codes that correspond to the color palettes found on the Texas Policy Lab website
\url{https://www.texaspolicylab.org/}
}
\concept{TPL palettes}
\keyword{datasets}
