% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_plot_test.R
\name{tpl_plot_test}
\alias{tpl_plot_test}
\title{Sample plots to test the appearance of \code{theme_TPL_print}}
\usage{
tpl_plot_test(type = "barplot", font = "adobe", include.logo = F)
}
\arguments{
\item{type}{Type of plot to be produced. Options include \code{barplot}, \code{scatterplot}, \code{boxplot}, \code{histogram}, and \code{Texas}.}

\item{font}{Font specification}

\item{include.logo}{Whether to include the TPL logo at the bottom of the plot}
}
\description{
Sample plots to test the appearance of \code{theme_TPL_print}
}
