% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tpl_logo.R
\name{add_tpl_logo}
\alias{add_tpl_logo}
\title{Texas Policy Lab \link{ggplot2} theme}
\usage{
add_tpl_logo(plot, position = "top right", scale = 1)
}
\arguments{
\item{plot}{The plot object passed to \code{add_tpl_logo}}

\item{position}{The position of the logo. Options are "top left", "top right", "bottom right", and "bottom left"}

\item{scale}{The scale of the TPL logo. Every single integer increase increases the scale multiplier (scale = 2 will double the size of the logo). Use decimals for fine-tuned scaling}
}
\description{
Add TPL logo to a single plot as grid object
Modified from https://gist.github.com/Ryo-N7/67ca1c364c342a82c4098918082ca445
}
