% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_continuous}
\alias{scale_fill_continuous}
\title{Continuous fill scale}
\usage{
scale_fill_continuous(palette = "sequential", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tpl_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{discrete_scale()}}
}
\description{
Continuous fill scale
}
