% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tpl_theme.R
\name{set_tpl_theme}
\alias{set_tpl_theme}
\title{The TPL \link{ggplot2} theme}
\usage{
set_tpl_theme(style = "print", font = "adobe")
}
\arguments{
\item{style}{The default theme style for the R session. Options are "print" or "map".}

\item{font}{The font for plot labels, axes, and titles. Options are "adobe" (Adobe Caslon Pro) and "lato" (Lato).}
}
\description{
\code{set_tpl_theme} provides a \link{ggplot2} theme formatted according to the
TPL website.
}
