% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tpl_print.R
\name{theme_tpl_print}
\alias{theme_tpl_print}
\title{A print \link{ggplot2} theme formatted in TPL style}
\usage{
theme_tpl_print(base_size = 11, base_family = "Adobe Caslon Pro",
  base_line_size = 0.5, base_rect_size = 0.5)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family. font options are Adobe Caslon Pro and Lato. Choices are "adobe" and "lato"}

\item{base_line_size}{base line}

\item{base_rect_size}{rectangle sizes}
}
\description{
\code{theme_tpl} provides a \link{ggplot2} theme formatted according to the TPL
website
}
