% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{Texas Policy Lab \link{ggplot2} theme}
\usage{
view_palette(palette = palette_tpl_main)
}
\arguments{
\item{palette}{A \code{palette_tpl_*} vector from \code{library(tpltheme)}.
Options are \code{palette_tpl_main}, \code{palette_tpl_diverging}, \code{palette_tpl_politics}, and \code{palette_tpl_sequential}}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for \code{palette_tpl_*} vectors.
These palettes were generated on www.coolors.co/.
}
\examples{
view_palette()
view_palette(palette_tpl_main)

}
